fix: la racion se recortaba de 2.5 kg a 100 g en silencio
Mario reporto que "el tiempo de accion de los motores es fijo en 5s" tras
cambiar los presets de la app a 2.5/4.5/5.5/6 kg. No era un tiempo fijo: era
un desbordamiento.

CmdPacket.grams era uint8_t y Estacion recortaba a 100. La app escribe
gramos de verdad — 2500, 4500, 5500, 6000 — asi que llegaban 100 g. Con
m_punto = 32.5 g/s eso son 3.18 s de sinfin mas 0.5 de pre-giro y 2 de cola:
5.7 s. Justo el "fijo en 5s" observado.

Se pedian 2.5 kg, salian ~100 g, y el sistema respondia ACK_OK. Un deficit
de 25 veces, silencioso, informando de exito. La clase de fallo que llevamos
todo el proyecto persiguiendo.

  grams pasa a uint16_t. El paquete crece de 10 a 11 B: ~5 ms mas de aire a
  SF9, nada frente a poder expresar lo que el sistema hace de verdad.
  GRAMS_MAX = 6000, atado al KG_MAX de la app.

  OBLIGA a reflashear LAS DOS placas a la vez: con una en 10 B y otra en 11,
  protoValidate rechaza todo por longitud.

Guarda de seguridad replanteada. Con dosis de kilos un ciclo legitimo dura
minutos, asi que un techo absoluto tardaria minutos en detectar una maquina
de estados colgada. Ahora el vigilante compara contra el t_on PLANIFICADO
mas DOSE_GUARD_MARGIN_MS: se detecta a los 5 s de pasarse. MAX_DOSE_MS
queda solo como techo de aceptacion, 240 s, que cubre los 185 s de 6 kg.

Lo que esto NO detecta es un sinfin atascado: seguiria girando su tiempo
completo. Haria falta leer la corriente por los pines IS del BTS7960, hoy
sin cablear. Anotado.

Verificado en placa: enlace intacto tras el cambio de tamaño, 0 descartados
y 0 perdidos.

PENDIENTE EN LA APP, anotado en CLAUDE.md: Aspersion.java ancla nivel 0 = 1 m
y el firmware ancla nivel 1 = 1 m reservando el 0 para apagar. Si el usuario
pide 1 m la app manda nivel 0 y el aspersor no gira. La formula correcta es
nivel = 1 + (metros - 1) / 1.3333.

+++ estacion/src/firebase_link.h
@@ -36,7 +36,7 @@
 /* Estado de los comandos, tal y como estan en la base ahora mismo. */
 typedef struct {
   NavCmd  nav;       /* resuelto desde las cuatro banderas    */
-  uint8_t grams;     /* /pwm, 0-100                           */
+  uint16_t grams;    /* /pwm, en gramos, hasta GRAMS_MAX        */
   uint8_t sprayer;   /* /aspersor, 0-10                       */
   bool    feed;      /* /motores, ya consumido si estaba true */
 } FbCommands;
